backend/requirements.txt
# Portail éducatif — dépendances back-end
Django>=5.1,<5.2
djangorestframework>=3.15,<4.0
djangorestframework-simplejwt>=5.3,<6.0
django-cors-headers>=4.4,<5.0

# Base de données
psycopg[binary]>=3.2,<4.0
dj-database-url>=2.2,<3.0

# Configuration via variables d'environnement
python-decouple>=3.8,<4.0

# IA / Claude
anthropic>=0.40,<1.0

# Outils de développement
ipython>=8.0
